Use yielding pytest_runtest_call() hook wrapper
Take advantage of `@pytest.mark.hookwrapper` facility and use more
idiomatic pytest_runtest_call(item) hook for performing capturing (in
fact, the same is used by the core 'pytest.capture' module).

Instead of messing around `report.longrepr` internals and since we
don't have a `report` instance at the moment of executing
pytest_runtest_call() anyway (it used to be created within
pytest_runtest_makereport() hook), use a higher-level
item.add_report_section() method to attach log output to a report.

However, there is a drawback that now reports show sections named like
'Captured stdlog ...' ('std' is hardcoded). Nevertheless this is likely
to be an issue of the pytest itself. It is addressed in
pytest-dev/pytest#1052, and that PR also contains a discussion on
deprecating Item.add_report_section(). However, the same approach is
still used by 'pytest.capture', so it should be fine to use it here as
well.

This change also removes the use of the deprecated '__multicall__', so
that fixes #3.