Replace CatchLogHandler with RecordingHandler that doesn't use streams
We only need formatted text (hence the stream) for log reporting and
for the caplog.text() method. On the other hand, we don't need a list
of records for the reporting facility.

So, we only leave storing of records in the RecordingHandler class, and
use ad hoc solutions using logging.StreamHandler where needed.